+++ src/app/fragment/[fragmentid]/edit/page.tsx
@@ -10,7 +10,7 @@ import {
   Container,
   TextArea,
 } from "@radix-ui/themes";
-import Header from "@/components/Header";
+import Header from "@/components/Organisims/Header";
 import { useState, useEffect, use } from "react";
 import { getTermFragment, updateTermFragment } from "@/functions/termFragments";
 import { TermFragment } from "@/types";
