+++ src/app/login/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { Box, Flex, Heading, Button, Container, Avatar, TextField, Text } from "@radix-ui/themes";
-import Header from "@/components/Header";
+import Header from "@/components/Organisims/Header";
 import { useAuth } from "@/contexts/AuthContext";
 import { useState, useEffect } from "react";
 import { useRouter } from "next/navigation";
