+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import Header from "@/components/Header";
+import Header from "@/components/Organisims/Header";
 import Link from "next/link";
 
 export default function Home() {
