+++ src/app/understood/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { Box, Flex, Heading, Button, Container, Card } from "@radix-ui/themes";
-import Header from "@/components/Header";
+import Header from "@/components/Organisims/Header";
 import Link from "next/link";
 import { useState, useEffect } from "react";
 import { getUserUnderstoodRecords } from "@/functions/understandingService";
